+++ src/http/http-client.ts
@@ -1,9 +1,11 @@
+import { Observable } from "rxjs";
+
 export type HttpResponse = {
     status: number;
     body: string;
 }
 
-export interface HttpClient<ResponseBodyType = string> {
-    send(url: string, method: 'POST' | 'GET' | 'PUT' | 'PATCH' | 'DELETE', headers?: {[key: string]: string}): Promise<{status: number, body: ResponseBodyType}>;
-    sendWithBody(url: string, method: 'POST' | 'PUT' | 'PATCH', data: ResponseBodyType, headers?: {[key: string]: string}): Promise<{status: number, body: ResponseBodyType}>
+export interface HttpClient {
+    send(url: string, method: 'POST' | 'GET' | 'PUT' | 'PATCH' | 'DELETE', headers?: {[key: string]: string}): Observable<{status: number, body: string}>;
+    sendWithBody(url: string, method: 'POST' | 'PUT' | 'PATCH', data: string, headers?: {[key: string]: string}): Observable<{status: number, body: string}>
 }