+++ src/http/http-repository.ts
@@ -3,80 +3,127 @@ import { Id } from "../id";
 import { PersistedModel } from "../model-base";
 import { CreateRepositoryOperation, DeleteRepositoryOperation, GetAllRepositoryOperation, GetDetailsRepositoryOperation, UpdateRepositoryOperation } from "../repository-operations";
 import { HttpClient } from "./http-client";
+import { map, Observable } from "rxjs";
 
 export type RouteFnWithModelId<ModelType> = (id: Id<ModelType>) => string;
 
-export class JsonHttpApiClient implements HttpClient<object> {
-
-    constructor(private _httpClient: HttpClient<string>) { }
-
-    send(url: string, method: "POST" | "PUT" | "PATCH" | "GET" | "DELETE", headers?: { [key: string]: string; } | undefined): Promise<{ status: number; body: object; }> {
-        return this._httpClient.send(url, method, { ...headers, 'Content-type': 'application/json' }).then((response) => {
-            return { ...response, body: JSON.parse(response.body) };
-        });
-    }
-
-    sendWithBody(url: string, method: "POST" | "PUT" | "PATCH", data: object, headers?: { [key: string]: string; } | undefined): Promise<{ status: number; body: object; }> {
-        return this._httpClient.sendWithBody(url, method, JSON.stringify(data), { ...headers, 'Content-type': 'application/json' }).then((response) => {
-            return { ...response, body: JSON.parse(response.body) };
-        });
-    }
-}
+export type SerializeFn<Input, Output> = (i: Input) => Output;
 
 export class HttpCreateRepositoryOperation<ModelType extends {}> implements CreateRepositoryOperation<ModelType> {
-    constructor(private _httpClient: JsonHttpApiClient, private _endpoint: string) { }
+    constructor(
+        private _httpClient: HttpClient,
+        private _endpoint: string,
+        private _serializeModel: SerializeFn<ModelType, string>,
+        private _deserializeModel: SerializeFn<string, PersistedModel<ModelType>>
+    ) { }
 
-    create(model: ModelType): Promise<PersistedModel<ModelType>> {
-        return this._httpClient.sendWithBody(this._endpoint, 'POST', model).then((result) => {
-            return new PersistedModel<ModelType>((result.body as any).id, model)
-        });
+    // modelType -> string
+    // string -> PersistedModel
+
+    create(model: ModelType): Observable<PersistedModel<ModelType>> {
+        return this._httpClient.sendWithBody(this._endpoint, 'POST', this._serializeModel(model))
+            .pipe(
+                map(({ body }) => {
+                    return this._deserializeModel(body);
+                })
+            );
     }
 }
 
 export class HttpGetAllRepositoryOperation<ModelType extends {}, FilterType = {}> implements GetAllRepositoryOperation<ModelType, FilterType> {
-    constructor(private _httpClient: JsonHttpApiClient, private _endpoint: string) { }
-    getAll(filter: Partial<FilterType>): Promise<PersistedModel<ModelType>[]> {
-        return this._httpClient.send(this._endpoint, 'GET').then((result) => {
-            if (!Array.isArray(result.body)) throw new Error("No array returned");
+    constructor(private _httpClient: HttpClient, private _endpoint: string,
+        private _serializeFilter: SerializeFn<FilterType, string>,
+        private _deserializeResponse: SerializeFn<string, PersistedModel<ModelType>[]>,
+    ) { }
+    // fitlerType -> string
+    // string -> PersistedModel<T>
 
-            return result.body.map((item) => {
-                return {
-                    id: item.id,
-                    model: item,
-                } as PersistedModel<ModelType>;
+    getAll(filter: Partial<FilterType>): Observable<PersistedModel<ModelType>[]> {
+        // TODO support filter
+        
+        return this._httpClient.send(this._endpoint, 'GET').pipe(
+            map(({ body }) => {
+                const models: PersistedModel<ModelType>[] = this._deserializeResponse(body);
+                if (!Array.isArray(models)) throw new Error("No array returned");
+
+                return models;
             })
-        });
+        )
     }
 }
 
 export class HttpGetDetailsForRepositoryOperation<ModelType extends {}> implements GetDetailsRepositoryOperation<ModelType> {
-    constructor(private _httpClient: JsonHttpApiClient, private _endpoint: RouteFnWithModelId<ModelType>) { }
-    getDetailsFor(id: Id<ModelType>): Promise<PersistedModel<ModelType> | undefined> {
-        return this._httpClient.send(this._endpoint(id), 'GET').then((result) => {
-            return result.body == undefined ? undefined : new PersistedModel<ModelType>((result.body as any)?.id, result.body as any);
-        });
+    constructor(private _httpClient: HttpClient, private _endpoint: RouteFnWithModelId<ModelType>, private _deseriailzeResponse: SerializeFn<string, PersistedModel<ModelType> | undefined>) { }
+
+    // nodata -> nodata
+    // string -> PersistedModel<T>
+
+    getDetailsFor(id: Id<ModelType>): Observable<PersistedModel<ModelType> | undefined> {
+        return this._httpClient.send(this._endpoint(id), 'GET').pipe(map(({ body }) => {
+            return this._deseriailzeResponse(body);
+        }));
     }
 }
 
 export class HttpUpdateRepositoryOperation<ModelType extends {}> implements UpdateRepositoryOperation<ModelType> {
-    constructor(private _httpClient: JsonHttpApiClient, private _endpoint: RouteFnWithModelId<ModelType>) { }
-    update(id: Id<ModelType>, updatedModel: ModelType): Promise<PersistedModel<ModelType>> {
-        return this._httpClient.sendWithBody(this._endpoint(id), 'PUT', updatedModel as any).then((result: any) => {
-            return new PersistedModel<ModelType>((result.body as any)?.id, result.body as any);
-        });
+    // modelType -> string
+    // string -> persistedModel
+    constructor(private _httpClient: HttpClient, private _endpoint: RouteFnWithModelId<ModelType>, private _serializeModel: SerializeFn<ModelType, string>, private _deserializeModel: SerializeFn<string, PersistedModel<ModelType>>) { }
+    update(id: Id<ModelType>, updatedModel: ModelType): Observable<PersistedModel<ModelType>> {
+        return this._httpClient.sendWithBody(this._endpoint(id), 'PUT', this._serializeModel(updatedModel)).pipe(
+            map(({ body }) => this._deserializeModel(body))
+        )
     }
 
 }
 
 export class HttpDeleteRepositoryOperation<ModelType extends {}> implements DeleteRepositoryOperation<ModelType> {
-    constructor(private _httpClient: JsonHttpApiClient, private _endpoint: RouteFnWithModelId<ModelType>) { }
-    delete(id: Id<ModelType>): Promise<boolean> {
-        return this._httpClient.send(this._endpoint(id), 'DELETE').then((result) => {
-            return result.status == 200;
-        });
+    constructor(private _httpClient: HttpClient, private _endpoint: RouteFnWithModelId<ModelType>) { }
+    delete(id: Id<ModelType>): Observable<boolean> {
+        return this._httpClient.send(this._endpoint(id), 'DELETE').pipe(
+            map(({ status }) => {
+                return status === 200;
+            })
+        )
+    }
+}
+
+export class JsonSerializationAdapter<ModelType, FilterType> {
+
+    public deserializeStringToPersistedModel: SerializeFn<string, PersistedModel<ModelType>> = (input: string) => {
+        const data = JSON.parse(input);
+
+        if (!("id" in data)) {
+            throw new Error("Could not deserialize ")
+        }
+
+        return new PersistedModel(new Id<ModelType>(data.id), data);
     }
 
+    public serializeModelToString: SerializeFn<ModelType, string> = (input: ModelType) => {
+        return JSON.stringify(input);
+    }
+
+    public serializeFilterToString: SerializeFn<FilterType, string> = (input: FilterType) => {
+        return ""; // TODO
+    }
+    
+    public deserializeArrayStringToPersistedModelArray: SerializeFn<string, PersistedModel<ModelType>[]> = (input: string) => {
+        const data = JSON.parse(input);
+
+        if (!Array.isArray(data))
+            throw new Error("Array expected for GetAllOperation");
+
+        return data.map(model => {
+            if (model.id === undefined) {
+                throw new Error("Expected every model to have an id");
+            }
+
+            return new PersistedModel(new Id<ModelType>(model.id), model);
+        })
+    }
 }
+
 export class HttpRepository<ModelType extends {}, FilterType = {}> implements CrudRepository<ModelType, FilterType> {
 
     protected _createOperation: HttpCreateRepositoryOperation<ModelType>;
@@ -86,35 +133,35 @@ export class HttpRepository<ModelType extends {}, FilterType = {}> implements Cr
     protected _deleteOperation: HttpDeleteRepositoryOperation<ModelType>;
 
     constructor(
-        protected _httpClient: JsonHttpApiClient,
+        protected _httpClient: HttpClient,
         protected _routeDefinitions: {
             create: string,
             getAll: string,
             getDetailsFor: RouteFnWithModelId<ModelType>,
             update: RouteFnWithModelId<ModelType>,
             delete: RouteFnWithModelId<ModelType>,
-        }
+        },
+        protected _serializationAdapter: any = JsonSerializationAdapter<ModelType, FilterType>,
     ) {
-        this._createOperation = new HttpCreateRepositoryOperation<ModelType>(_httpClient, _routeDefinitions.create);
-        this._getAllOperation = new HttpGetAllRepositoryOperation<ModelType>(_httpClient, _routeDefinitions.getAll);
-        this._getDetailsForOperation = new HttpGetDetailsForRepositoryOperation<ModelType>(_httpClient, _routeDefinitions.getDetailsFor);
-        this._updateOperation = new HttpUpdateRepositoryOperation<ModelType>(_httpClient, _routeDefinitions.update);
+        this._createOperation = new HttpCreateRepositoryOperation<ModelType>(_httpClient, _routeDefinitions.create, this._serializationAdapter.serializeModelToString, this._serializationAdapter.deserializeStringToPersistedModel);
+        this._getAllOperation = new HttpGetAllRepositoryOperation<ModelType, FilterType>(_httpClient, _routeDefinitions.getAll, this._serializationAdapter.serializeFilterToString, this._serializationAdapter.deserializeArrayStringToPersistedModelArray);
+        this._getDetailsForOperation = new HttpGetDetailsForRepositoryOperation<ModelType>(_httpClient, _routeDefinitions.getDetailsFor, this._serializationAdapter.deserializeStringToPersistedModel);
+        this._updateOperation = new HttpUpdateRepositoryOperation<ModelType>(_httpClient, _routeDefinitions.update, this._serializationAdapter.serializeModelToString, this._serializationAdapter.deserializeStringToPersistedModel);
         this._deleteOperation = new HttpDeleteRepositoryOperation<ModelType>(_httpClient, _routeDefinitions.delete);
     }
-    create(model: ModelType): Promise<PersistedModel<ModelType>> {
+    create(model: ModelType): Observable<PersistedModel<ModelType>> {
         return this._createOperation.create(model);
     }
-    getAll(filter: Partial<FilterType>): Promise<PersistedModel<ModelType>[]> {
+    getAll(filter: Partial<FilterType>): Observable<PersistedModel<ModelType>[]> {
         return this._getAllOperation.getAll(filter);
     }
-    getDetailsFor(id: Id<ModelType>): Promise<PersistedModel<ModelType> | undefined> {
+    getDetailsFor(id: Id<ModelType>): Observable<PersistedModel<ModelType> | undefined> {
         return this._getDetailsForOperation.getDetailsFor(id);
     }
-    update(id: Id<ModelType>, updatePayload: ModelType): Promise<PersistedModel<ModelType>> {
+    update(id: Id<ModelType>, updatePayload: ModelType): Observable<PersistedModel<ModelType>> {
         return this._updateOperation.update(id, updatePayload);
     }
-    delete(id: Id<ModelType>): Promise<boolean> {
+    delete(id: Id<ModelType>): Observable<boolean> {
         return this._deleteOperation.delete(id);
     }
-
 }