+++ src/local-storage/local-storage-storage-adapter.ts
@@ -1,3 +1,4 @@
+import { Observable, of } from "rxjs";
 import { StorageAdapter } from "../in-memory-repository";
 import { PersistedModel } from "../model-base";
 
@@ -5,13 +6,14 @@ export class LocalStorageStorageAdapter implements StorageAdapter {
     
     constructor(protected _key: string) { }
 
-    setItems<T>(value: T[]): Promise<boolean> {
+    setItems<T>(value: T[]): Observable<boolean> {
         localStorage.setItem(this._key, JSON.stringify(value));
-        return Promise.resolve<boolean>(true);
+        return of(true);
     }
-    getItems<T>(): Promise<PersistedModel<T>[]> {
-        const stringFromLocalStorage = localStorage.getItem(this._key);
-        return stringFromLocalStorage === null ? [] : JSON.parse(stringFromLocalStorage);
+    getItems<T>(): Observable<PersistedModel<T>[]> {
+        const data = localStorage.getItem(this._key)
+        
+        return of(data === null ? [] : JSON.parse(data));
     }
 
 }