+++ src/repository-operations/create-repository-operation.ts
@@ -1,5 +1,6 @@
+import { Observable } from "rxjs";
 import { PersistedModel } from "../model-base";
 
 export interface CreateRepositoryOperation<ModelType> {
-    create(model: ModelType): Promise<PersistedModel<ModelType>>;
+    create(model: ModelType): Observable<PersistedModel<ModelType>>;
 }