+++ src/repository-operations/delete-repository-operation.ts
@@ -1,5 +1,6 @@
+import { Observable } from "rxjs";
 import { Id } from "../id";
 
 export interface DeleteRepositoryOperation<ModelType> {
-    delete(id: Id<ModelType>): Promise<boolean>;
+    delete(id: Id<ModelType>): Observable<boolean>;
 }