+++ src/repository-operations/get-all-repository-operation.ts
@@ -1,5 +1,6 @@
+import { Observable } from "rxjs";
 import { PersistedModel } from "../model-base";
 
 export interface GetAllRepositoryOperation<ModelType, FilterType> {
-    getAll(filter: Partial<FilterType>): Promise<PersistedModel<ModelType>[]>;
+    getAll(filter: Partial<FilterType>): Observable<PersistedModel<ModelType>[]>;
 }