+++ src/repository-operations/get-details-repository-operation.ts
@@ -1,6 +1,7 @@
+import { Observable } from "rxjs";
 import { Id } from "../id";
 import { PersistedModel } from "../model-base";
 
 export interface GetDetailsRepositoryOperation<ModelType> {
-    getDetailsFor(id: Id<ModelType>): Promise<PersistedModel<ModelType> | undefined>;
+    getDetailsFor(id: Id<ModelType>): Observable<PersistedModel<ModelType> | undefined>;
 }