+++ src/repository-operations/update-repository-operation.ts
@@ -1,6 +1,7 @@
+import { Observable } from "rxjs";
 import { Id } from "../id";
 import { PersistedModel } from "../model-base";
 
 export interface UpdateRepositoryOperation<ModelType> {
-    update(id: Id<ModelType>, updatedModel: ModelType): Promise<PersistedModel<ModelType>>;
+    update(id: Id<ModelType>, updatedModel: ModelType): Observable<PersistedModel<ModelType>>;
 }